==> 1. Nasdaq stock screener.ipynb <==
# Import af selenium og webdriver
from selenium import webdriver
from webdriver_manager.chrome import ChromeDriverManager

# Importing webdriverwait
from selenium.webdriver.support.ui import WebDriverWait

# To use the expected_conditions
from selenium.webdriver.support import expected_conditions as EC

# To use the function "By"
from selenium.webdriver.common.by import By

# To use the ActionChains 
from selenium.webdriver.common.action_chains import ActionChains

# import os
import os

# Importing time
import time

# Please read below

This is the first file in order to get all the data collection, processing, analysis and visualization. 

PLEASE notice that there is a great ammount of data to be downloaded from the internet through web scraping and API's. This will take several days to download.

All data was eventually converted to csv files. All these files can be downloaded from our githup (https://github.com/sebasfaurby/Social23Group12.git) along with all the code.

Also, notice that all codes use the same folder as current working directory.


# Change the download directory to the current working directory
download_directory = os.getcwd()

chrome_options = webdriver.ChromeOptions()
prefs = {'download.default_directory': download_directory}
chrome_options.add_experimental_option('prefs', prefs)

driver = webdriver.Chrome(options=chrome_options)

# URL
url = 'https://www.nasdaq.com/market-activity/stocks/screener'

# Set up the WebDriver (provide the path to your driver executable)
# driver = webdriver.Chrome()

# Navigate to the webpage
driver.get(url) 
# maximize window using webdriver
driver.maximize_window()

#### Clicking cookie button ####

try:
    # Wait for the button to be clickable
    button_cookie = WebDriverWait(driver, 10).until(
        EC.element_to_be_clickable((By.ID, "onetrust-accept-btn-handler"))
    )
    # Click the button
    
    button_cookie.click()
except Exception as e:
    print("An error occurred:", str(e))

# Filtering for the NASDAQ exchange
button = driver.find_element(By.ID, "radioItemNASDAQ")
button.click()

# Filtering for market cap
market_caps = ['mega','large','mid','small','micro'] 
for market_cap in market_caps:
    button_market_cap = driver.find_element(By.XPATH, f"//label[@for='checkboxItem{market_cap}']")
    button_market_cap.click()
    time.sleep(1)
    
# How much to scroll down
scroll_increment = 900

# Scroll down
driver.execute_script(f"window.scrollBy(0, {scroll_increment});")

# Find the scroll down menu element by class name
scroll_menu = driver.find_element(By.CLASS_NAME, 'nasdaq-screener__check-box--country-item')

# Scroll to the bottom of the menu
driver.execute_script("arguments[0].scrollTop = arguments[0].scrollHeight", scroll_menu)
time.sleep(1)

# Check the box for United States
checkbox_item = driver.find_element(By.XPATH, '//label[@for="checkboxItemunited_states"]')
checkbox_item.click()

# Apply filters to Nasdaq stock screener
button_apply = driver.find_element(By.CLASS_NAME, 'nasdaq-screener__form-button--apply')
button_apply.click()
time.sleep(1)

# Download the csv file with the applied filters
button_csv = driver.find_element(By.CLASS_NAME, "nasdaq-screener__form-button--download")
button_csv.click()
