==> 5. Data processing.ipynb <==
# Data processing

### Intro
This code will get the files with information of each stock and do following:
- Compute some technical indicators for each stock. Hence mfi, rsi, P/E and historical market cap
- Join all stocks together in a large datafile

import tqdm
import time
import os
import tqdm
import datetime
import matplotlib.pyplot as plt
import numpy as np
import pandas as pd
from pathlib import Path

eod_path = Path.cwd()/"Eod" #path to create folder where end of day stock prices will be stored
fundamentals_path = Path.cwd()/"Fundamentals"
SingleStock_data = Path.cwd()/"SingleStock_data"

cwd = Path.cwd()

### Calculate MFI

#First to functions to compute gains and losses will be needed
def gain(x):
    return ((x > 0) * x).sum()


def loss(x):
    return ((x < 0) * x).sum()


def mfi(df, n=14):
    high = df["high"]
    low = df["low"]
    close = df["adjusted_close"]
    volume = df["volume"]
    
    typical_price = (high + low + close)/3
    money_flow = typical_price * volume
    
    mf_sign = np.where(typical_price > typical_price.shift(1), 1, -1)
    signed_mf = money_flow * mf_sign
    
    mf_avg_gain = signed_mf.rolling(n).apply(gain, raw=True)
    mf_avg_loss = signed_mf.rolling(n).apply(loss, raw=True)
    
    mfi = (100 - (100 / (1 + (mf_avg_gain / abs(mf_avg_loss))))).to_numpy()
    
    return mfi


### Calculate volatility

def vol(df, period = 21):
    daily_return = df['adjusted_close'].pct_change()
    
    volatility = daily_return.rolling(window = period).std()
    
    return volatility


def rsi(df, periods = 14):
    
    delta = df['adjusted_close'].diff()

    # Make two series: one for lower closes and one for higher closes
    up = delta.clip(lower=0)
    down = -1 * delta.clip(upper=0)
    
    ma_up = up.ewm(com = periods - 1, adjust=True, min_periods = periods).mean()
    ma_down = down.ewm(com = periods - 1, adjust=True, min_periods = periods).mean()

    rsi = ma_up / ma_down
    rsi = 100 - (100/(1 + rsi))
    return rsi

def gain(x):
    return ((x > 0) * x).sum()


def loss(x):
    return ((x < 0) * x).sum()


def mfi(df, n=14):
    high = df["high"]
    low = df["low"]
    close = df["adjusted_close"]
    volume = df["volume"]
    
    typical_price = (high + low + close)/3
    money_flow = typical_price * volume
    
    mf_sign = np.where(typical_price > typical_price.shift(1), 1, -1)
    signed_mf = money_flow * mf_sign
    
    mf_avg_gain = signed_mf.rolling(n).apply(gain, raw=True)
    mf_avg_loss = signed_mf.rolling(n).apply(loss, raw=True)
    
    mfi = (100 - (100 / (1 + (mf_avg_gain / abs(mf_avg_loss))))).to_numpy()
    
    return mfi

def vol(df, period = 21):
    daily_return = df['adjusted_close'].pct_change()
    
    volatility = daily_return.rolling(window = period).std()
    
    return volatility

def compute_technical(df):
    df["date"] =pd.to_datetime(df["date"])
    #compute RSI
    df['RSI'] = rsi(df)
    
    #compute MFI
    df["MFI"] = mfi(df)
    
    #compute volatility
    df["volatility"]=vol(df)
    
    return df
    
    

## Create master dataframe:
Time to join all dataframes into one

def compute_data(df):
    
    df["date"] = pd.to_datetime(df["date"])
    df = compute_technical(df)
    df = df.drop('close',axis =1)
    df["returns"] = (df["adjusted_close"]-df["adjusted_close"].shift(-1))/df["adjusted_close"].shift(-1)*100
    df['prev_returns'] = df['returns'].shift(-1)
    df['returns_pred'] = df['returns'].shift(1)
    df = df[(df["date"]>='2022-06-01')&(df["date"]<='2023-06-02')]
    
    return df

def master_df(master, df, ticker):
    """
    Finds data for a single stocks and joins it to masterdataframe on=date how=left
    
    Args(pandas.DataFrame): dataframe of the company that will be joined to the master dataframe
                            the existing master dataframe
    Args(str): ticker of the company that will be joined to master dataframe
    
    returns the updated master dataframe
    """

    df["date"] = pd.to_datetime(df["date"])
    df["Ticker"] = ticker

    master = pd.concat([master,df])
    master["date"] = pd.to_datetime(master["date"])
    master = master.sort_values(["date", "Ticker"], ascending=[False, True]).reset_index(drop = True)
    
    return master


def compute_fundamentals(df):
    df["date"] = pd.to_datetime(df["date"])
    df["marketcap"] = df["adjusted_close"]*df["outstandingShares"]
    df["PE"] = df["adjusted_close"]/df["epsActual"]
    #RSI missing
    #MFI missing
    #Standard deviation missing
    
    return df

#get tickers for stocks
# tickers = pd.read_csv("Tickers_final.csv")["Ticker"].values
tickers = pd.read_csv("TickerNames_1956.csv")['0'].values
tickers

### Master dataframe for EoD only

#delete this sentence # delete this if you want to rund the code - its just a fail safe

master_eod = pd.DataFrame() #define empty dataframe. THis is the final dataframe
for ticker in tqdm.tqdm(tickers):
    #get datafram from csv
    df = pd.read_csv(eod_path/f"{ticker}.csv")
    master_eod = master_df(master_eod,df,ticker)
    master_eod.to_csv("Master_eod.csv", index = False)
    

master_eod = master_eod.sort_values("date", ascending=False)

mast = pd.read_csv("Master_eod.csv").drop("Unnamed: 0", axis =1)
mast["date"]=pd.to_datetime(mast["date"])
mast.sort_values(['date', 'Ticker'], ascending=[False,True])

### Master dataframe with technicals, returns and taget returns

delete this sentence # delete this sentence if you are sure you want to run this code

master_tech = pd.DataFrame() #define empty dataframe. THis is the final dataframe

for ticker in tqdm.tqdm(tickers):
    #get datafram from csv
    df = pd.read_csv(eod_path/f"{ticker}.csv")
    df = compute_data(df)
    master_tech = master_df(master_tech, df, ticker)
    master_tech.to_csv("Master_with_technicals_news.csv", index = False)

master_tech = pd.read_csv("master_df.csv")
master_tech = master_tech.sort_values(["date","Ticker"], ascending = [False,True])

master_tech

# Find the location of NaN values
nan_locations = np.where(pd.isnull(master_tech))

# The 'nan_locations' tuple contains two arrays: one for row indices and one for column indices
row_loc, cols_loc = nan_locations

# after analysing NaN values we notice that the missings values for MFI column are simply 0
# The calculation returns NaN when the high or low price is the same for more than 14 days
#as the average becomes 0 and we end up dividing by zero. Therefore, all NaN in MFI column are filled by zero

master_tech["MFI"].fillna(0, inplace = True)

#For volatility there were only 4 companies missing the last two days of volatility data
#as we wont use the 2nd, we simply fill these values with 
master_tech["volatility"].fillna(method = 'bfill', inplace = True)
master_tech.to_csv("Master_with_technicals_news.csv", index = False)
